Light Bake, Blockout, Post Processing
Did another light bake, with new doors from the blockout. Also adjusted the conveyor belt position again, as it was clipping with a metal beam. Changed the post-processing so it doesnt affect certain layer objects such as the player and the enemy, this is adjustable. Also added a stripe caution texture to the bottom of the doors temporarily to see how it looks.